requirements.txt
requests~=2.22.0
scipy~=1.1.0
matplotlib~=3.1.1
mpmath~=1.1.0
sympy~=1.4
discord.py>=1.0.1
aiohttp~=3.4.4
numpy~=1.16.4
Pillow~=6.2.0
dice~=2.4.2
hsluv~=0.0.2
numba~=0.45.1
pint~=0.9
regex~=2019.08.19
tensorflow==1.15.2
